requirements.txt
# Dependencias originales
pandas==3.0.0
requests==2.31.0
beautifulsoup4==4.14.3
soupsieve==2.8.3
lxml==5.1.0

# Flask y extensiones
Flask==3.0.0
Flask-CORS==4.0.0

# Firebase
firebase-admin==6.4.0

# Variables de entorno
python-dotenv==1.0.0

# Seguridad
bcrypt==4.1.2